01_Health_and_wellness_planner/pyproject.toml
[project]
name = "01-health-and-wellness-planner"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "openai-agents>=0.1.0",
    "pydantic>=2.11.7",
    "streamlit>=1.46.1",
]
